Add a migration to remove the deprecated old LastServedNotice and CurrentPrivacyPreference tables. These were replaced by the LastServedNoticeV2 and CurrentPrivacyPreferenceV2 tables in January 2024.
- Remove the deprecated models from "fides.api.models.privacy_preference".
- Move the replacement models from "privacy_preference_v2" to "privacy_preference" now that there is only one set.
- Rename test_privacy_preference_v2 to test_privacy_preference
