+++ src/fides/api/db/base.py
@@ -34,10 +34,6 @@ from fides.api.models.privacy_notice import (
     PrivacyNoticeTemplate,
 )
 from fides.api.models.privacy_preference import (
-    DeprecatedCurrentPrivacyPreference,
-    DeprecatedLastServedNotice,
-)
-from fides.api.models.privacy_preference_v2 import (
     CurrentPrivacyPreferenceV2,
     LastServedNoticeV2,
     PrivacyPreferenceHistory,
