+++ tests/fixtures/application_fixtures.py
@@ -49,12 +49,12 @@ from fides.api.models.privacy_notice import (
     PrivacyNoticeRegion,
     PrivacyNoticeTemplate,
 )
-from fides.api.models.privacy_preference import ServingComponent
-from fides.api.models.privacy_preference_v2 import (
+from fides.api.models.privacy_preference import (
     ConsentIdentitiesMixin,
     CurrentPrivacyPreferenceV2,
     PrivacyPreferenceHistory,
     ServedNoticeHistory,
+    ServingComponent,
 )
 from fides.api.models.privacy_request import (
     Consent,
